% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_H2O_PMM.R
\name{binary.H2O.PMM}
\alias{binary.H2O.PMM}
\title{Function to generate imputations using PMM and H2O for data with a binary outcome}
\usage{
binary.H2O.PMM(Y, X, newdata, SL.library, k)
}
\arguments{
\item{Y}{blah}

\item{X}{blah}

\item{newdata}{blah}

\item{SL.library}{blah}

\item{k}{blah}
}
\value{
nothing
}
\description{
Function to generate imputations using PMM and H2O for data with a binary outcome
}
