% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_H2O_regression.R
\name{binary.H2O.regression}
\alias{binary.H2O.regression}
\title{Function to generate imputations using regression and H2O for data with a binary outcome}
\usage{
binary.H2O.regression(Y, X, newdata, SL.library)
}
\arguments{
\item{Y}{blah}

\item{X}{blah}

\item{newdata}{blah}

\item{SL.library}{blah}
}
\value{
nothing
}
\description{
Function to generate imputations using regression and H2O for data with a binary outcome
}
