% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_SuperLearner_PMM.R
\name{binary.SuperLearner.PMM}
\alias{binary.SuperLearner.PMM}
\title{Function to generate imputations using PMM and SuperLearner for data with a binary outcome}
\usage{
binary.SuperLearner.PMM(Y, X, wy, SL.library, k, ...)
}
\arguments{
\item{Y}{blah}

\item{X}{blah}

\item{SL.library}{blah}

\item{k}{blah}

\item{...}{further arguments passed to SuperLearner}

\item{newdata}{blah}
}
\value{
nothing
}
\description{
Function to generate imputations using PMM and SuperLearner for data with a binary outcome
}
