% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_H2O_regression.R
\name{continuous.H2O.norm}
\alias{continuous.H2O.norm}
\title{Function to generate imputations using regression and H2O for data with a continuous outcome}
\usage{
continuous.H2O.norm(Y, X, newdata, SL.library)
}
\arguments{
\item{Y}{blah}

\item{X}{blah}

\item{newdata}{blah}

\item{SL.library}{blah}
}
\value{
nothing
}
\description{
Function to generate imputations using regression and H2O for data with a continuous outcome
}
