% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_H2O_PMM.R
\name{continuous.H2O.PMM}
\alias{continuous.H2O.PMM}
\title{Function to generate imputations using PMM and H2O for data with a continuous outcome}
\usage{
continuous.H2O.PMM(Y, X, newdata, SL.library, k)
}
\arguments{
\item{Y}{blah}

\item{X}{blah}

\item{newdata}{blah}

\item{SL.library}{blah}

\item{k}{blah}
}
\value{
nothing
}
\description{
Function to generate imputations using PMM and H2O for data with a continuous outcome
}
