% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_H2O_regression.R
\name{continuous.H2O}
\alias{continuous.H2O}
\title{Function to generate imputations using regression and H2O for data with a continuous outcome}
\usage{
continuous.H2O(Y, X, newdata, SL.library)
}
\arguments{
\item{Y}{Vector of observed values of the variable to be imputed.}

\item{X}{Numeric matrix of variables to be used as predictors in H2O methods
with rows corresponding to observed values of the variable to be imputed.}

\item{newdata}{Numeric matrix of variables to as predictors in H2O methods
with rows corresponding to missing values of the variable to be imputed. The
H2O model makes predictions from this matrix to determine the
imputation-generating distribution.}

\item{SL.library}{List of functions and parameters to be passed on to the
H2O function}
}
\value{
Numeric vector of randomly generated imputed values.
}
\description{
Function to generate imputations using regression and H2O for data with a continuous outcome
}
