% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_SuperLearner_regression.R
\name{continuous.SuperLearner.norm}
\alias{continuous.SuperLearner.norm}
\title{Function to generate imputations using regression and SuperLearner for data with a continuous outcome}
\usage{
continuous.SuperLearner.norm(y, x, wy, SL.library, SL.CV, method.weights, ...)
}
\arguments{
\item{SL.library}{blah}

\item{SL.CV}{blah}

\item{method.weights}{blah}

\item{...}{further arguments passed to SuperLearner}

\item{Y}{blah}

\item{X}{blah}

\item{newdata}{blah}
}
\value{
nothing
}
\description{
Function to generate imputations using regression and SuperLearner for data with a continuous outcome
}
