% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_SuperLearner_PMM.R
\name{continuous.SuperLearner.PMM}
\alias{continuous.SuperLearner.PMM}
\title{Function to generate imputations using PMM and SuperLearner for data with a continuous outcome}
\usage{
continuous.SuperLearner.PMM(Y, X, newdata, SL.library, k, ...)
}
\arguments{
\item{Y}{blah}

\item{X}{blah}

\item{newdata}{blah}

\item{SL.library}{blah}

\item{k}{blah}

\item{...}{further arguments passed to SuperLearner}
}
\value{
nothing
}
\description{
Function to generate imputations using PMM and SuperLearner for data with a continuous outcome
}
