% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_SuperLearner_regression.R
\name{continuous.SuperLearner.regression}
\alias{continuous.SuperLearner.regression}
\title{Function to generate imputations using regression and SuperLearner for data with a continuous outcome}
\usage{
continuous.SuperLearner.regression(Y, X, newdata, SL.library, SL.CV, ...)
}
\arguments{
\item{Y}{blah}

\item{X}{blah}

\item{newdata}{blah}

\item{SL.library}{blah}

\item{SL.CV}{blah}

\item{...}{further arguments passed to SuperLearner}
}
\value{
nothing
}
\description{
Function to generate imputations using regression and SuperLearner for data with a continuous outcome
}
