% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_SuperLearner_regression.R
\name{continuousSuperLearner}
\alias{continuousSuperLearner}
\title{Function to generate imputations using regression and SuperLearner for data with a continuous outcome}
\usage{
continuousSuperLearner(
  y,
  x,
  wy,
  SL.library,
  kernel,
  bw,
  bw.update,
  lambda,
  ...
)
}
\arguments{
\item{y}{Vector of observed values of the variable to be imputed.}

\item{x}{Numeric matrix of variables to be used as predictors in H2O methods
with rows corresponding to observed values of the variable to be imputed.}

\item{wy}{Logical vector of length length(y). A TRUE value indicates
locations in y for which imputations are created.}

\item{SL.library}{Either a character vector of prediction algorithms or a
list containing character vectors. A list of functions included in the
SuperLearner package can be found with SuperLearner::listWrappers().}

\item{kernel}{One of "gaussian",...  Kernel function used to compute weights.}

\item{bw}{NULL or numeric value for bandwidth of kernel function (as standard deviations of the kernel).}

\item{bw.update}{eriogjaoier}

\item{lambda}{NULL or numeric value for bandwidth for kernel (as half-width of the kernel).}

\item{...}{further arguments passed to SuperLearner.}
}
\value{
Numeric vector of randomly drawn imputed values.
}
\description{
Function to generate imputations using regression and SuperLearner for data with a continuous outcome
}
