% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{gaussianKernel}
\alias{gaussianKernel}
\title{Kernel functions used for local imputation}
\usage{
gaussianKernel(x, xcenter, bw = 1, lambda = NULL)
}
\arguments{
\item{x}{blah}

\item{xcenter}{blah}

\item{bw}{blah}

\item{lambda}{blah}
}
\value{
nothing
}
\description{
Kernel functions used for local imputation
}
