% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife_bandwidth_selection.R
\name{jackknife.bandwidth.selection}
\alias{jackknife.bandwidth.selection}
\title{Jackknife method for selection bandwidth}
\usage{
jackknife.bandwidth.selection(
  i,
  bwGrid,
  preds,
  y,
  delta,
  lambda = NULL,
  imputation,
  kernel,
  weights
)
}
\arguments{
\item{i}{blah}

\item{bwGrid}{blah}

\item{preds}{blah}

\item{y}{blah}

\item{delta}{blah}

\item{lambda}{blah}

\item{imputation}{blah}

\item{kernel}{blah}

\item{weights}{blah}
}
\value{
bandwidth
}
\description{
Jackknife method for selection bandwidth
}
