% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife_variance.R
\name{jackknifeVariance}
\alias{jackknifeVariance}
\title{Computes jackknife variance}
\usage{
jackknifeVariance(j, kernMatrix, n, m, delta, y)
}
\arguments{
\item{j}{blah}

\item{kernMatrix}{blah}

\item{n}{blah}

\item{m}{blah}

\item{delta}{blah}

\item{y}{blah}
}
\value{
variance
}
\description{
Computes jackknife variance
}
