% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnImpute.R
\name{knnImpute}
\alias{knnImpute}
\title{Function to generate imputations using PMM and SuperLearner for data with a continuous outcome}
\usage{
knnImpute(x, candidates, k)
}
\arguments{
\item{x}{blah}

\item{candidates}{blah}

\item{k}{blah}
}
\value{
nothing
}
\description{
Function to generate imputations using PMM and SuperLearner for data with a continuous outcome
}
