% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_impute_h2o_pmm.R
\name{mice.impute.h2o.pmm}
\alias{mice.impute.h2o.pmm}
\title{Predictive Mean Matching (PMM) imputation method for \code{mice} using Super Learner via \code{h2o}.}
\usage{
mice.impute.h2o.pmm(y, ry, x, wy = NULL, h2o.models, donors = 5)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length length(y) indicating the the subset y[ry]
of elements in y to which the imputation model is fitted. The ry generally
distinguishes the observed (TRUE) and missing values (FALSE) in y.}

\item{x}{Numeric design matrix with length(y) rows with predictors for y.
Matrix x may have no missing values.}

\item{wy}{Logical vector of length length(y). A TRUE value indicates
locations in y for which imputations are created.}

\item{h2o.models}{A named list of character vectors specifying
prediction algorithms and arguments to be passed to h2o.  See details below
for examples on the structure.}

\item{donors}{The size of the donor pool among which a draw is made. The
default is donors = 5L. Setting donors = 1L always selects the closest match,
but is not recommended. Values between 3L and 10L provide the best results
in most cases (Morris et al, 2015).}

\item{...}{Further arguments passed to \code{SuperLearner} or \code{h2o}.}
}
\value{
Vector with imputed data, same type as y, and of length sum(wy)
}
\description{
Samples imputations for univariate missing data using PMM.
Distances for PMM are computed using predictions from a Super Learner model
instead of the standard regression model.
}
\details{
This method is similar to what van Buuren (2018) describes as type 0
matching. Distances are determined only from predicted values from a Super
Learner model.  The closest \code{k} predictions are used to determine the
donor pool.

The backend fitting the Super Learner models is H2O.  Models are specified
as a list of named character vectors.  The first element of each vector is
the name of the function being called in H2O.
}
\examples{
\dontrun{
# Define the h2o models
h2o.mods = list(TO DO)

# Run mice using the h2o normal method
imps = mice::mice(mice::nhanes, m = 5, method = "h2o.norm",
                   h2o.models = h2o.mods)

# list the actual imputations
imp$imp$bmi

# first completed data matrix
complete(imps)

# predict hypertensive from age, bmi, and total serum cholesterol

}

}
