% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_impute_SuperLearner_pmm.R
\name{mice.impute.SuperLearner.pmm}
\alias{mice.impute.SuperLearner.pmm}
\title{Predictive Mean Matching (PMM) imputation method for \code{mice} using Super Learner via \code{SuperLearner}.}
\usage{
mice.impute.SuperLearner.pmm(y, ry, x, wy = NULL, SL.library, donors = 5, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length length(y) indicating the the subset y[ry]
of elements in y to which the imputation model is fitted. The ry generally
distinguishes the observed (TRUE) and missing values (FALSE) in y.}

\item{x}{Numeric design matrix with length(y) rows with predictors for y.
Matrix x may have no missing values.}

\item{wy}{Logical vector of length length(y). A TRUE value indicates
locations in y for which imputations are created.}

\item{SL.library}{A character vector of prediction algorithms or list
containing character vectors as specified by the SuperLearner package.}

\item{donors}{The size of the donor pool among which a draw is made. The
default is donors = 5L. Setting donors = 1L always selects the closest match,
but is not recommended. Values between 3L and 10L provide the best results
in most cases (Morris et al, 2015).}

\item{...}{Further arguments passed to \code{SuperLearner}.}
}
\value{
Vector with imputed data, same type as y, and of length sum(wy)
}
\description{
Samples imputations for univariate missing data using PMM.
Distances for PMM are computed using predictions from a Super Learner model
instead of the standard regression model.
}
