% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{gaussianKernel}
\alias{gaussianKernel}
\title{Kernel functions used for local imputation}
\usage{
gaussianKernel(x, xcenter, bw = 1, lambda = NULL)
}
\arguments{
\item{x}{numeric vector of values to weight.}

\item{xcenter}{numeric value to center the kernel.}

\item{bw}{bandwidth of the kernel.}

\item{lambda}{kernel radius, function of \code{bw}.}
}
\value{
kernel values for \code{x} centered at \code{xcenter}.
}
\description{
Kernel functions used for local imputation
}
