% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife_bandwidth_selection.R
\name{jackknifeBandwidthSelection}
\alias{jackknifeBandwidthSelection}
\title{Jackknife method for selection bandwidth}
\usage{
jackknifeBandwidthSelection(i, bwGrid, preds, y, delta, lambda = NULL, kernel)
}
\arguments{
\item{i}{blah}

\item{bwGrid}{blah}

\item{preds}{blah}

\item{y}{blah}

\item{delta}{blah}

\item{lambda}{blah}

\item{kernel}{blah}
}
\value{
bandwidth
}
\description{
Jackknife method for selection bandwidth
}
