% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_imputaiton.R
\name{localImputation}
\alias{localImputation}
\title{Function to generate imputations using regression and SuperLearner for data with a continuous outcome}
\usage{
localImputation(
  i,
  x,
  y,
  delta,
  bw,
  imputation = c("semiparametric", "nonparametric"),
  kernel = c("gaussian"),
  weightType = c("nadaraya-watson")
)
}
\arguments{
\item{x}{blah}

\item{y}{blah}

\item{delta}{blah}

\item{bw}{blah}

\item{imputation}{blah}

\item{kernel}{further arguments passed to SuperLearner}

\item{weightType}{blah}

\item{preds}{blah}
}
\value{
nothing
}
\description{
Function to generate imputations using regression and SuperLearner for data with a continuous outcome
}
