% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_imputaiton.R
\name{localImputation}
\alias{localImputation}
\title{Function to generate imputations using non-parametric and semi-parametric local imputation methods.}
\usage{
localImputation(
  i,
  preds,
  y,
  delta,
  bw = NULL,
  lambda = NULL,
  kernel = c("gaussian", "uniform", "triangular")
)
}
\arguments{
\item{i}{Integer referring to the index of the missing value to be imputed.}

\item{preds}{Numeric vector of predictions of missing values from SuperLearner or H2O.}

\item{y}{Numeric vector for variable to be imputed.}

\item{delta}{Binary vector of length length(y) with 1 where y is observed and 0 where y is missing.}

\item{bw}{NULL or numeric value for bandwidth of kernel function (as standard deviations of the kernel).}

\item{lambda}{NULL or numeric value for bandwidth for kernel (as half-width of the kernel).}

\item{kernel}{Kernel function used to compute weights.}
}
\value{
Numeric vector of randomly drawn imputed values.
}
\description{
Function to generate imputations using non-parametric and semi-parametric local imputation methods.
}
