% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_impute_h2o.R
\name{mice.impute.h2o.norm}
\alias{mice.impute.h2o.norm}
\title{Normal model imputation method for \code{mice} parameterized by Super Learner via \code{h2o}.}
\usage{
mice.impute.h2o.norm(y, ry, x, wy = NULL, h2o.models, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length length(y) indicating the the subset y[ry]
of elements in y to which the imputation model is fitted. The ry generally
distinguishes the observed (TRUE) and missing values (FALSE) in y.}

\item{x}{Numeric design matrix with length(y) rows with predictors for y.
Matrix x may have no missing values.}

\item{wy}{Logical vector of length length(y). A TRUE value indicates
locations in y for which imputations are created.}

\item{h2o.models}{A named list of character vectors specifying
prediction algorithms and arguments to be passed to h2o.  See details below
for examples on the structure.}

\item{...}{Further arguments passed to \code{SuperLearner} or \code{h2o}.}
}
\value{
Vector with imputed data, same type as y, and of length sum(wy)
}
\description{
Calculates imputations for univariate missing data by the normal model.
Parameters for the normal distribution generating the data are determined
using a SuperLearner model fit in H2O.
}
\details{
TO DO: Add details
}
\examples{
\dontrun{
# Define the h2o models
h2o.mods = list()

# Run mice using the h2o normal method
imps = mice::mice(mice::nhanes, m = 5, method = "h2o.norm",
                   h2o.models = h2o.mods)

# list the actual imputations
imp$imp$bmi

# first completed data matrix
complete(imps)

# predict hypertensive from age, bmi, and total serum cholesterol

}

}
